Gracefully handle control FD death when inheriting process group.
We cannot rely on killpg if we don't own the process group.
Refactor out shutdown_processes() instead and handle things similar to
SIGINT/SIGTERM.

+++ cli/fossilize_replay_linux.hpp
@@ -704,7 +704,7 @@ static void update_target_running_processes(std::vector<ProcessProgress> &child_
 		size_t num_processes = child_processes.size();
 		for (size_t i = 0; i < num_processes && to_stop; i++)
 		{
-			if (child_processes[i].pid >= 0 && !child_processes[i].stopped)
+			if (child_processes[i].pid >= 0 && !child_processes[i].stopped && !child_processes[i].expect_kill)
 			{
 				if (::kill(child_processes[i].pid, SIGSTOP) == 0)
 				{
@@ -722,7 +722,7 @@ static void update_target_running_processes(std::vector<ProcessProgress> &child_
 		size_t num_processes = child_processes.size();
 		for (size_t i = 0; i < num_processes && to_wake_up; i++)
 		{
-			if (child_processes[i].pid > 0 && child_processes[i].stopped)
+			if (child_processes[i].pid > 0 && child_processes[i].stopped && !child_processes[i].expect_kill)
 			{
 				if (::kill(child_processes[i].pid, SIGCONT) == 0)
 				{
@@ -758,7 +758,20 @@ static void handle_control_command(const char *command)
 		LOGE("Unrecognized control command: %s.\n", command);
 }
 
-static void handle_control_commands()
+static void shutdown_processes(std::vector<ProcessProgress> &child_processes)
+{
+	for (auto &child_process : child_processes)
+	{
+		if (child_process.pid >= 0)
+		{
+			child_process.stopped = false;
+			child_process.expect_kill = true;
+			kill(child_process.pid, SIGKILL);
+		}
+	}
+}
+
+static void handle_control_commands(std::vector<ProcessProgress> &child_processes)
 {
 	char buffer[4096];
 	ssize_t ret;
@@ -775,12 +788,7 @@ static void handle_control_commands()
 		if (Global::control_fd_is_sentinel)
 		{
 			LOGI("Parent process closed control FD with no notification, terminating early ...\n");
-			// control FD in parent process was hung, kill every child process and ourselves and then exit.
-			if (killpg(getpid(), SIGKILL) < 0)
-			{
-				LOGE("Failed to kill process group ... This should not happen.\n");
-				abort();
-			}
+			shutdown_processes(child_processes);
 		}
 		close_fd = true;
 	}
@@ -1150,21 +1158,12 @@ static int run_master_process(const VulkanDevice::Options &opts,
 						// someone explicitly SIGKILLs it, so make sure we take down our children
 						// before we go.
 						LOGI("Received signal %s, cleaning up ...\n", info.ssi_signo == SIGINT ? "SIGINT" : "SIGTERM");
-
-						for (auto &child_process : child_processes)
-						{
-							if (child_process.pid >= 0)
-							{
-								child_process.stopped = false;
-								child_process.expect_kill = true;
-								kill(child_process.pid, SIGKILL);
-							}
-						}
+						shutdown_processes(child_processes);
 					}
 				}
 				else if (e.data.u32 == POLL_VALUE_CONTROL_FD)
 				{
-					handle_control_commands();
+					handle_control_commands(child_processes);
 					// After a control command treat it as having received a timer event
 					// so we react quickly to requested changes.
 					update_target_running_processes(child_processes);
